Before:
```julia
julia> const x = randn(Xoshiro(1337), 10^4);

julia> @benchmark bayesian_blocks(x, resolution=100.0, min_counts=2)
BenchmarkTools.Trial: 6 samples with 1 evaluation.
 Range (min … max):  911.158 ms … 916.188 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     912.652 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   913.097 ms ±   1.914 ms  ┊ GC (mean ± σ):  0.00% ± 0.00%

  █   █       █          █               █                    █
  █▁▁▁█▁▁▁▁▁▁▁█▁▁▁▁▁▁▁▁▁▁█▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█ ▁
  911 ms           Histogram: frequency by time          916 ms <

 Memory estimate: 705.55 KiB, allocs estimate: 36.
```

After:
```julia
julia> @benchmark bayesian_blocks(x, resolution=100.0, min_counts=2)
BenchmarkTools.Trial: 6 samples with 1 evaluation.
 Range (min … max):  896.677 ms … 901.405 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     899.464 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   899.281 ms ±   1.522 ms  ┊ GC (mean ± σ):  0.00% ± 0.00%

  █                             █   █  ██                     █
  █▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█▁▁▁█▁▁██▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█ ▁
  897 ms           Histogram: frequency by time          901 ms <

 Memory estimate: 548.83 KiB, allocs estimate: 30.
```