@Moelf 2022-10-21
actually the speed up doesn't matter as much because now I realized this doesn't scale beyond 100k points.

But the type annotation fix should be useful, since sometimes people have integer data