@francescoalemanno 2022-10-21
thanks a lot Jerry! I've read issue #2 , I think tomorrow I will work on it :)